+++ tests/testthat/test-cnd-abort.R
@@ -666,3 +666,64 @@ test_that("call can be a quosure or contain quosures", {
   err <- catch_error(abort("foo", call = quo(f(!!quo(g())))))
   expect_equal(err$call, quote(f(g())))
 })
+
+test_that("`parent = NA` signals a non-chained rethrow", {
+  local_options(
+    rlang_trace_top_env = current_env(),
+    rlang_trace_format_srcrefs = FALSE
+  )
+
+  ff <- function() gg()
+  gg <- function() hh()
+
+  foo <- function() bar()
+  bar <- function() baz()
+  baz <- function() stop("bar")
+
+  expect_snapshot({
+    "Absent parent causes bad trace bottom"
+    hh <- function() {
+      withCallingHandlers(foo(), error = function(cnd) {
+        abort(cnd_header(cnd))
+      })
+    }
+    print(err(ff()))
+
+    "Missing parent allows correct trace bottom"
+    hh <- function() {
+      withCallingHandlers(foo(), error = function(cnd) {
+        abort(cnd_header(cnd), parent = NA)
+      })
+    }
+    print(err(ff()))
+
+    "Wrapped handler"
+    handler1 <- function(cnd, call = caller_env()) handler2(cnd, call)
+    handler2 <- function(cnd, call) abort(cnd_header(cnd), parent = NA, call = call)
+    hh <- function() {
+      withCallingHandlers(
+        foo(),
+        error = function(cnd) handler1(cnd)
+      )
+    }
+    print(err(ff()))
+
+    "Wrapped handler, `try_fetch()`"
+    hh <- function() {
+      try_fetch(
+        foo(),
+        error = function(cnd) handler1(cnd)
+      )
+    }
+    print(err(ff()))
+
+    "Wrapped handler, incorrect `call`"
+    hh <- function() {
+      withCallingHandlers(
+        foo(),
+        error = handler1
+      )
+    }
+    print(err(ff()))
+  })
+})